Fix PHP TAPTestEngine no-op `continue` in switch.
Summary:
`arc unit` failed locally with the following error: `Error while loading file
".../tap_test_engine/src/TAPTestEngine.php": "continue" targeting switch is
equivalent to "break". Did you mean to use "continue 2"?`

It's almost certainly a difference in local setup. However, it seems the
intention was to escape from the `foreach` but a plain `default: continue` in a
`switch` is effectively a no-op.